User can now make the form on top of other...
windows. This can be done by clicking the View>Always On Top menubar
button. There's a hidden debug label that will show the user the status
of Always On Top. Make sure you comment out the
debugLabelForAlwaysOnTop.Hide() code in the Form1_Load or MainForm_Load
event.

At runtime this label will populate with a value. The value will either
be 0False or 1True, where the 1 or 0 is the CheckedState of the
menubarAlwaysOnTopButton and True or False is the Boolean value of the
corresponding My.Settings value.