Making the Always On Top feature a little better.
The way I did this is by telling aaformMainWindow to not be on top when
the options window opens. However, this would lead to the main window
saying that it's set to be on top when it's really not. Thus, when the
options window closes, the main window is set to be on top again if the
My.Settings.alwaysOnTop is set to True. If the main form is already set
to not be on top, then the options window will make sure that it's not
on top to avoid any problems later on.