MenuBar BackColor is no longer needed...
I now set colorMenubarBackColor to colorMenuItemBackColor after setting
colorMenuItemBackColor to the MenuItem BackColor XML element. This fixes
a bug which occurred if there wasn't anything to set MenuBar BackColor
to so the menubar wouldn't update any of it except for the menu items
when the user mouses over the items.