Hide the "Revert to Default Theme"...
button if the user disabled the theme engine. Fixes #92.

I'm still disabling the button to ensure no theme-related changes occur
when the theme engine is disabled.

As a result of this change, the tooltip text for the button has been
reworded. Of particular interest is the text that says that the button
can only be used if the theme engine is enabled in the Options window.
After all, if someone doesn't want to use the theme engine, why have a
useless (to them) button cluttering the UI?