Properly open Options window from Quickmenu even...
when main window is set to Always On Top. This is done by placing "Me"
into the forceOptionsWindowTab.ShowDialog() code ("Me" goes inside the
parenthesis). Fixes #53 partially.

This only partially fixes #53 because duplicate Options windows show up
when I open the Options window from the Quickmenu. At least the Options
window can be accessed easily without having to right-click on the
Quickmenu icon if the Options window goes behind other windows now.

The "Always On Top" feature is also impacted by this change as the
Options window now keeps both itself and the main window on top of all
other windows when "Always On Top" is enabled. Previously, opening the
Options window would temporarily disable "Always On Top" for the main
window and the Options window then re-enable the feature when the
Options window closed. Thus, "Always On Top" should now work a bit
closer to "always" and further from "sometimes;" up until now it was
more like "Sometimes On Top."