About window respects proper "Always On Top"...
functioning. I fixed this issue in
94edb9d042fb9e06c9244cda54ff83ae7bd6ab03 for the Options window, but I
forgot to check the About window at the time and it turns out that the
About window had trouble with "Always On Top" as well. However, both the
Options window and the About window now work with "Always On Top" almost
all the time, with the one exception of the ability to open multiple
Options windows from the Quickmenu as referenced in the last few
comments of #53.