Allow themes to specify which version of...
the theme engine to use in case something works better in an older
version for a particular theme. This isn't tested yet, but this works
with an XML element at the level of the Description, Name, and Author
elements in the XML theme files. The element this uses is
"UseThemeEngineVersion". By default, if the element is missing, then
"1.0" will be assumed.