Improve theme engine version specification.
Added debug info on which version of the theme engine a theme wants to
use to main window. It's not in the best place, but it works for now. It
should be moved to its own label, though.

Additionally, if the version of the theme engine to be used is not
specified in the theme file, it is assumed that the theme wants to use
version 1.01. If the version that is specified is less than 1.01, even
if it's specified as 1.0, then it is assumed that the theme wants to use
version 1.01. If the theme file specifies a version larger than or equal
to 1.01, the version of the theme engine that the theme file specifies
will be used.

Also, the version of the theme engine to be used will no longer be
printed into the Visual Studio Immediate Window.

Any changes made to the UXL Launcher Theme Engine from now until when
version 3.1 is released will keep the theme engine at version 1.01. If
the theme engine is modified at all between the stable release of UXL
Launcher 3.1 and UXL Launcher 3.2, the theme engine version number will
be incremented by 0.01, that is, the version number will increase from
1.01 to 1.02 if there is any work done on the theme engine in that time.
This version number will be displayed in the About window below the
information for UXL Launcher itself, similar to how HideSettingsPages
and hsp_registry-helper information is presented in the
HideSettingsPages About window.