Implement most of "Hide When Minimized"...
as specified in #47, but there still needs to be a menuitem under the
"View" menu for it to enable and disable the feature whenever the user
wants. This can only be done on a monitor at 96 DPI (dots per inch)
without messing up the layout, so the easiest way for me to do it is to
do it on my desktop computer. This branch won't be merged into master
until that is done.