Debug labels cleanup commit...
This commit is mainly to cleanup debug labels on the main form. There
was a new one added for the "UseThemeEngineVersion" XML element, the
debug labels for "Always On Top" were moved to the rightmost column,
descriptions for each of the theme debug labels were added to the code
that updates the debug labels, and an unneeded line of code was deleted.