Fixed known issue in pull request #108.
I fixed this by adding the code "Me.Hide" after the main window is
hidden from the taskbar and placing "Me.Show" right before the main
window is shown in the taskbar when either double-clicking on the
notification icon or clicking "Show UXL Launcher" in the Quickmenu.

Now, the main window doesn't show up in the Windows Alt+Tab interface if
"Hide When Minimized" is enabled and it's minimized.