Improve show UXL Launcher functionality...
provided in the Quickmenu. Now, if the main window is behind another
window, it'll be brought to the front when the user clicks the "Show UXL
Launcher" button in the Quickmenu. This needs to be copied to the
double-click action as well, or the code needs to be in another sub not
in two places at once.