Rename the "FileNotFound_CustomTheme" exception...
to "UXLLauncher.ThemeEngine.FileNotFound_CustomTheme" so that
developers, designers, and users know that it's not a system exception.
This one isn't actually an exception set up like regular Windows
exceptions, but just a reference so that there's a string that can be
easily referred to online or in documentation. Also, the output for the
StatusLabel BorderSides code has been rearranged.