Output theme user chose when clicking...
the "Default Theme" button again. This broke a while back due to theme
engine changes and I'm just now fixing it. It doesn't look as good as it
used to, but it's better than nothing. I still need to change some other
code. Also, it only outputs if
My.Settings.debugmodeShowThemeEngineOutput is set to True.