Now, menuitem colors can be different from...
the menubar color. The only problem is that there's no button background
area anymore. That will hopefully be fixed soon. Some code used from
http://www.vbforums.com/showthread.php?539578-Custom-VisualStudio2008-style-MenuStrip-and-ToolStrip-Renderer&p=3333808&viewfull=1#post3333808

Some other changes in this commit include using a solid brush for the
menubar background and a new property for the "UXLToolstripRenderer"
known as "DropdownBackColor". This color is used to change the color of
the dropdown boxes in the menubar.