Make sure the menuitem button arrow is visible...
on dark backgrounds when the mouse cursor isn't on a menuitem that has
an arrow. This uses the theme file's MenuItem ForeColor property so that
the buttons are the same color as the menuitem text when menuitems
aren't highlighted.

Additionally, add some comments and only set e.TextColor and not that
plus e.Item.ForeColor when the item text color gets changed by
highlighting the item.