For now, the "Experience - Blue" theme should...
be complete. Some things don't look right, but that's a limitation of
the theme engine and I can change things in the future to look better.
One thing that looks off that's not a color is the status bar. I didn't
make the status bar borders themselves customizable from themes, but
this can be fixed in a future version of the theme engine.

I also used slightly different colors without just sampling from the
Windows XP themes to not make Microsoft unhappy.