Finish error message for custom themes not...
being allowed. If you as an organization or business want to use this
feature, please note that it's not 100% perfect and if it's enabled in
the configuration file then the configuration file is saved while UXL
Launcher is open, custom themes can still be loaded with the theme debug
buttons.