If a theme name such as "Maudern" is pasted into...
the theme list dropdown box, the theme info box on the left will update
with the theme's info. For now, this requires that the theme name
matches the file in My.Resources exactly, though this will be fixed
shortly. Also, to test this out, the DropDown style will need to be
selected in the Visual Studio Designer properties pane for the dropdown
box to paste text into it.