Catch some exceptions.
These exceptions being caught now allow the user to type an exact theme
name into the theme list in the Options window, but that functionality
is only for testing for now. I'm on a HiDPI device, so I won't push my
changes to the Options window, but you can set the theme list combobox
DropDownStyle property to DropDown (Windows Forms default) to try it
out.