Info for custom themes now works properly...
for the most part. There's some text that gets displayed if the custom
theme couldn't be found, and that needs to be updated to also ask the
user to make sure the file in the custom theme path textbox exists.

Also, the code for the theme list when it's (Custom) needs to go into
another sub so that it can be called by both the TextChanged property on
the theme list combobox and the custom theme path textbox TextChanged
event. This makes sure the theme info is updating when needed.