Increase About window banner resolution.
It's now 512x1024 as opposed to 256x512 previously, including in version
3.0. This change means that the logo banner in the About window should
look cleaner and higher resolution than in previous versions. The actual
logo at the top of the banner was imported as an SVG file from the
regular logo's Photoshop file before being exported as a PNG file.

The borders have also been updated as well.