Hiding the Options window seems not necessary...
anymore since the Options window gets closed when settings are saved,
but I'm still keeping it in the code just in case. If it really isn't
needed, it'll be removed, but quite a while ago, there was a bug where
duplicate Options windows would open if the user chose to open the
Options window to change their preferences to make sure settings are
configured correctly. That was issue #46.