Begin to remove the MSI Install Method...
checkbox and its functionality from the app. See also #100. For now,
I'll just switch over the Office 2013 option in the Options window to
not enable the MSI Install Method checkbox, so that the Office 365/C2R
checkbox is the one that'll be used.

Some benefits of this change include not having an extra setting in the
Config file to deal with and being able to just see if
My.Settings.userHasOfficeThreeSixFive = False instead of comparing
Office version numbers and this setting with the other MSI Install
Method setting.

This would also make updating the app easier and remove duplicate code.