Add File>Open... button.
See also #67. For now, this is just a menuitem that has the text
"Open..." and a shortcut of Ctrl+O.

A separator has also been added between Open and Exit.

I'm probably going to organize the File menu similar to LibreOffice
version 4.4.3.2. This would be as follows:

File
- New
- Microsoft Word document (Ctrl+N)
- Microsoft Excel workbook
- Microsoft PowerPoint presentation
- Microsoft Outlook email
- Open... (Ctrl+O)
- Recent files
- (recent files here)
- Privacy options... (button to open Options window Privacy tab to turn
off showing recent files and an option to change how many recent files
are shown)
- (separator)
- Exit (Alt+F4)