Instead of seeing if the value for...
My.Settings.userOfficeVersion is "16" or "14" when creating the full
configured path if the user has Office 365, only see if the
userOfficeVersion setting IS NOT "15".

This makes it so that, if the user manually sets the Office version in
the config file and has Office 365 compatibility turned on, the apps
should still start properly. This should also work if future versions of
Office as long as Microsoft doesn't change the overall install location
after Office 2016 (excluding the version, like "Office14").

Also, the configured location for Office 2013 via Office 365 has the
"root" part all lowercase; it used to be "Root".

See also #119.

I still need to make the Office version dropdown in the Options window
show "Office(version number)" for Office versions manually set in the
config file that aren't officially supported.

If the user then saves their settings with that "Office(version number)"
text in the dropdown, just save the current value of
My.Settings.userOfficeVersion back to itself.

The dropdown will need something like an
Items.Add("Office" & My.Settings.userOfficeVersion) thing.

Also update a comment.