If the user's version of Office isn't officially...
supported by UXL Launcher, show that version at the bottom of the Office
version dropdown box.

See also #119.

I still need to make it so that saving settings with the unsupported
version of Office just saves the current My.Settings.userOfficeVersion
value.

This feature works by seeing if a String resource of
My.Resources.supportedOfficeVersionList contains the value of
My.Settings.userOfficeVersion in the config file.

If that version number isn't in that list, there will be an extra item
in the Office version dropdown as "Office(version number)" at the bottom
of the list.

However, if the version number is contained, that extra entry won't be
included.