Unsupported Office versions are now...
saved when clicking the "Save" button in the Options window if that's
what is listed in the Office version selector.

This might not be perfect, and Office versions that contain parts of
"14,15,16" might not display in the Office version dropdown due to how
this feature works, but this should work well enough as long as Office
10 or higher is in the config file for My.Settings.userOfficeVersion.

Fixes #119.

It's nice that, at least for now, that bug is fixed. If Microsoft
releases a preview of Office 2019 before Version 3.2 is out, this
feature should allow it to work as long as Microsoft continues to use
the current overall folder path of "C:\Program Files (x86)\Microsoft
Office\root\Office(version numer)".