Main window now uses a TableLayoutPanel to make it...
resizeable. As a result, the theme engine has been changed to version 2.01 instead of 1.03 due to changes that could break if someone just changes the theme engine compatibility version to 1.03. Adding theming to the Options and About windows also seem like they should be a new, major version rather than just a minor version.

Changes in this commit:
- FlowLayoutPanel is now a TableLayoutPanel to make resizing the main window possible if the user wants. Maximizing isn't possible yet, but that may be enabled as well.
- ThemeEngine has a new "MainWindowTableLayoutPanel" entry for XML files to modify. This is so that the TableLayoutPanel colors in the Options and About windows aren't used for the main window.
- If a theme is set to use a version older than 2.01, then the old behavior of applying the FlowLayoutPanel's color to the main window will be used. Otherwise, the table layout panel's color will be pulled from "MainWindowTableLayoutPanel" as this is the preferred way to do it in TE 2.01. This is in place to prevent major theme file breakage.