Themes can specify if they want the tablelayoutpanel...
color to apply in the About window's About tab where there's a tablelayoutpanel control. This doesn't yet apply the tabcontrol color instead if the theme doesn't want it, so that needs to be added.

Also when applying a theme not supporting TE1.03, the About window needs to apply Window for that control.