Allow themes to specify "LiteralNothing" so that the...
theme engine uses "Nothing" as a color for the statusbar back color. This makes it so that the statusbar looks like it would if the theme engine were turned off.

Still needs to have it only work with themes that are compatible with TE1.03 or greater.

Also Button1 now just shows what the statusbar back color is.