Add MaudernClassic as a theme option in case anyone...
preferred the colors of that version. There are a few additions to make it compatible with TE1.03 fully, but the general color scheme is the same as it was in Maudern v1.0.

Also copy some code into a place it should have been but forgot to do. Changed all the theme files to be "Content" for the build action as well.