Simplify usage of portable theme engine.
What's changed is that now there's a LoadTheme sub that has the theme file passed from the host application, the form that the host application wants it to be applied to, and, optionally, a boolean value to specify if a theme is a filename or not. By default, it'll be assumed that a theme file input is a filename, so it'll look under .\Themes\(theme filename from theme input) for it. This needs to make sure the theme file exists before trying to load it, though.

After the theme file is loaded, it's sent to the ApplyTheme sub along with the form it's meant to apply to.

Additionally, host applications no longer need to specify a toolstrip pro renderer in their own code, as it's set up in the theme engine instead. All that's needed now is for the host application to asign the Renderer property to the toolstrips using the one in the theme engine.