Move theme info groupbox to bottom + hide it if...
the theme engine is disabled. Also remove a bunch of code that was used for showing and hiding the debug labels, but it's not needed anymore since now it's all on another form and the entry point (debug menu) alone can be hidden.

A little extra spacing has been added below some of the labels in the theme info groupbox as well so that the text isn't clipped.