Hide/show theme info groupbox as needed + disable...
buttons in groupbox if the theme engine is disabled. That groupbox is hidden if the theme engine is disabled on startup, but if it's enabled on startup, it'll be hidden  when the debug window is re-opened if the theme engine is disabled at runtime.

Some comments have been added and old code from when the theme tester buttons were on the main form has been removed.