Merge GetAttributeSafe and GetInnerTextSafe together...
into GetPropertySafe. This allows choosing between using InnerText or Attributes by just changing a boolean between False or True, as True is the default and gets the Attribute from GetAttribute. False will get the InnerText from GetInnerText.

Additionally, extra forward slashes on the end of the node are now removed. This would be possible in case the Description tag is looked for by using GetPropertySafe("Description", "", "", False, True). Even if that node doesn't exist, removing the forward slash is necessary or else Xpath will complain that it's invalid.