Allow the user to restart the application if the theme engine...
is turned on or off. This'll prevent the "test settings" messagebox from showing up if they do restart, but that's intended since there could be other things that would influence the results at a later point. If the user clicks "No", they won't be bothered to restart for the rest of the session.

Fixes #177.