Add label saying to pick a control to theme from the...
dropdown above if an unsupported control is picked, or more usually if the combobox is empty, such as on startup. Should this default to "Button"? Also move code from the handler for combobox text updating to its own sub so we can call it on startup and display the message label. Update copyright years, too.